+++ {{ cookiecutter.project_name }}/composition_repository/djangocms_audio/settings.py
@@ -1,0 +1,15 @@
+from project_composer.marker import EnabledApplicationMarker
+
+
+class DjangoCMSAudioSettings(EnabledApplicationMarker):
+    DJANGOCMS_AUDIO_ALLOWED_EXTENSIONS = ["mp3", "ogg", "wav"]
+
+    @classmethod
+    def setup(cls):
+        super().setup()
+
+        cls.INSTALLED_APPS.extend(
+            [
+                "djangocms_audio",
+            ]
+        )