+++ {{ cookiecutter.project_name }}/composition_repository/djangocms_file/__init__.py
@@ -1,0 +1,4 @@
+DEPENDENCIES = [
+    "djangocms",
+    "django_filer",
+]